--- README.old.md
@@ -1,2 +1,0 @@
-# food-delivery-front-end
-Food delivery website Front-end part
