+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "food-delivery-front-end",
   "version": "0.1.0",
-  "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
