+++ src/components/App.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const App = () => (
+    <div className="App">
+        <p>Test</p>
+    </div>
+);
+
+export default App;
